misc/HtmlTestSuite/results/test-0957.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>scheduler: external files added through DOM should not block further parsing while loading</title>
        <script src="/resources/testharness.js"/>
        <script src="/resources/testharnessreport.js"/>
        <script src="testlib/testlib.js"/>
        <script>
                testlib.addScript('',{src:'scripts/find-body.js?pipe=trickle(d1)'},document.getElementsByTagName('head')[0], true );
        </script>
</head>
<body>
        <script>
                testlib.addScript('', {src:'scripts/find-foo.js?pipe=trickle(d1)'}, document.getElementsByTagName('head')[0], true);
        </script>
        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <p><span id="foo"/></p>

        <script type="text/javascript">
        var t = async_test()

        function test() {
                if(!(window.findFooLoaded &amp;&amp; window.findBodyLoaded)) {
                    return setTimeout(t.step_func(test), 200);
                }
                assert_any(assert_array_equals, eventOrder,
                           [['document.body: &lt;BODY&gt;', 'found #foo element: YES'],
                            ['found #foo element: YES', 'document.body: &lt;BODY&gt;']]);
                t.done();
        }
        onload = t.step_func(test)
        </script>

</body></html>
